chore: tune default resource limits and pool sizes in docker-compose
Reduce SYNC_DB_POOL from 5 to 3, DATABASE_POOL_MAX_CONNECTIONS from 8 to 5, and adjust vector embedding batch sizes (32->16) and lease time (1800->600). Remove POLL_INTERVAL_SECONDS and VECTOR_EMBEDDING_MAX_CONCURRENT_COMPLETES environment variables to streamline configuration.

+++ services/vec-worker/src/text_builder.rs
@@ -896,9 +896,9 @@ async fn build_baseline_profile(
         SELECT
             bssid,
             metric,
-            p5,
-            p50,
-            p95,
+            p5::float8,
+            p50::float8,
+            p95::float8,
             updated_at
         FROM vec_baseline_profiles
         WHERE bssid = $1
@@ -931,9 +931,9 @@ async fn build_baseline_profiles_batch(
             bssid AS query_key,
             bssid,
             metric,
-            p5,
-            p50,
-            p95,
+            p5::float8,
+            p50::float8,
+            p95::float8,
             updated_at
         FROM vec_baseline_profiles
         WHERE bssid = ANY($1::text[])
